Removed type hints, externalized node
